move button styling to theme and remove legacy Button component
