+++ src/components/App/App.tsx
@@ -4,23 +4,7 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import { EditSession, Header } from "../index";
 import Timer from "../Timer/timer";
 import styles from "./App.module.scss";
-
-const theme = createMuiTheme({
-  palette: {
-    type: "dark",
-    primary: {
-      main: "#ffd803",
-      contrastText: "#272343",
-    },
-    secondary: {
-      main: "#f25f4c",
-      contrastText: "#272343",
-    },
-  },
-  typography: {
-    fontFamily: "Work Sans, Roboto, Helvetica, Arial, sans-serif",
-  },
-});
+import theme from "../../styles/theme";
 
 const App: React.FC = () => {
   return (
