--- src/components/Button/Button.module.scss
@@ -1,41 +1,0 @@
-@use "src/styles/tokens" as *;
-
-@mixin button {
-  background-color: $card-background;
-  border: none;
-  border-radius: 3px;
-  color: $on-card;
-  display: inline-block;
-  font-size: $font-size-l;
-  font-variant: all-small-caps;
-  font-weight: 300;
-  letter-spacing: 0.5px;
-  line-height: 1.2;
-  padding: $s-size;
-  text-decoration: none;
-  cursor: pointer;
-  &:disabled {
-    cursor: not-allowed;
-    opacity: 0.5;
-  }
-}
-
-.primary {
-  @include button;
-  background-color: $primary;
-  color: $on-primary;
-}
-
-.primary-on-primary {
-  @extend .secondary;
-}
-
-.secondary {
-  @include button;
-  background-color: $secondary;
-  color: $on-secondary;
-}
-
-.secondary-on-primary {
-  @include button;
-}
