--- src/components/Button/index.ts
@@ -1,3 +1,0 @@
-import StyledButton from "./Button";
-
-export default StyledButton;
