--- src/components/Button/test/Button.test.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import "@testing-library/jest-dom";
-import { render } from "@testing-library/react";
-import Button from "../Button";
-
-describe("Button", () => {
-  it("renders without error", () => {});
-});
