+++ src/components/index.ts
@@ -7,7 +7,6 @@ import App from "./App";
 import ProjectSearch from "./Inputs/Input";
 import Card from "./Card";
 import Header from "./Header";
-import StyledButton from "./Button";
 
 export {
   /* PLOP_INJECT_EXPORT */
@@ -20,5 +19,4 @@ export {
   ProjectSearch,
   Card,
   Header,
-  StyledButton,
 };
